+++ injectable_generator/lib/generators/injectable_config_generator.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-import 'dart:io';
 
 import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
@@ -7,19 +6,15 @@ import 'package:code_builder/code_builder.dart';
 import 'package:collection/collection.dart';
 import 'package:dart_style/dart_style.dart';
 import 'package:glob/glob.dart';
-import 'package:glob/list_local_fs.dart';
 import 'package:injectable/injectable.dart';
 import 'package:injectable_generator/code_builder/allocator.dart';
 import 'package:injectable_generator/code_builder/library_builder.dart';
 import 'package:injectable_generator/models/dependency_config.dart';
 import 'package:injectable_generator/models/external_module_config.dart';
-import 'package:injectable_generator/models/importable_type.dart';
 import 'package:injectable_generator/resolver_utils.dart';
 import 'package:injectable_generator/resolvers/importable_type_resolver.dart';
 import 'package:recase/recase.dart';
 import 'package:source_gen/source_gen.dart';
-
-import 'package:injectable_generator/utils.dart';
 
 import '../resolvers/utils.dart' show throwIf;
 
@@ -40,7 +35,6 @@ class InjectableConfigGenerator extends GeneratorForAnnotation<InjectableInit> {
         .listValue
         .map((e) => e.toStringValue());
 
-    final usesNullSafety = annotation.read('usesNullSafety').boolValue;
     final isMicroPackage = annotation.read('_isMicroPackage').boolValue;
     final usesConstructorCallback = annotation
         .read('usesConstructorCallback')
@@ -57,16 +51,11 @@ class InjectableConfigGenerator extends GeneratorForAnnotation<InjectableInit> {
         .setValue
         .map((e) => e.getField('name')?.toStringValue());
 
-    final includeMicroPackages = annotation
-        .read("includeMicroPackages")
-        .boolValue;
-
     final generateAccessors = annotation.read("generateAccessors").boolValue;
 
-    final allowMultipleRegistrations =
-        annotation.read('allowMultipleRegistrations').boolValue;
-
-    final rootDir = annotation.peek('rootDir')?.stringValue;
+    final allowMultipleRegistrations = annotation
+        .read('allowMultipleRegistrations')
+        .boolValue;
 
     final dirPattern = generateForDir.length > 1
         ? '{${generateForDir.join(',')}}'
@@ -120,27 +109,6 @@ class InjectableConfigGenerator extends GeneratorForAnnotation<InjectableInit> {
     final microPackagesModules = microPackageModulesBefore.union(
       microPackageModulesAfter,
     );
-    if (!isMicroPackage && includeMicroPackages) {
-      final glob = Glob('**.module.dart', recursive: true);
-      final filesStream = glob.list(root: rootDir);
-
-      await for (final match in filesStream) {
-        final commentAnnotation = File(match.path).readAsLinesSync().first;
-        if (commentAnnotation.contains('@GeneratedMicroModule')) {
-          final segments = commentAnnotation.split(';');
-          if (segments.length == 3) {
-            final externalModule = ExternalModuleConfig(
-              ImportableType(name: segments[1], import: segments[2]),
-            );
-            if (!microPackagesModules.any(
-              (e) => externalModule.module == e.module,
-            )) {
-              microPackageModulesBefore.add(externalModule);
-            }
-          }
-        }
-      }
-    }
 
     final ignoreTypesInPackages = annotation
         .read('ignoreUnregisteredTypesInPackages')
@@ -210,7 +178,6 @@ class InjectableConfigGenerator extends GeneratorForAnnotation<InjectableInit> {
     final emitter = DartEmitter(
       allocator: HashedAllocator(),
       orderDirectives: true,
-      useNullSafetySyntax: usesNullSafety,
     );
 
     final output = DartFormatter(
@@ -221,11 +188,7 @@ class InjectableConfigGenerator extends GeneratorForAnnotation<InjectableInit> {
       final outputId = buildStep.inputId.changeExtension('.module.dart');
       return buildStep.writeAsString(
         outputId,
-        [
-          '//@GeneratedMicroModule;${capitalize(buildStep.inputId.package.pascalCase)}PackageModule;${outputId.uri}',
-          defaultFileHeader,
-          output,
-        ].join('\n'),
+        [defaultFileHeader, output].join('\n'),
       );
     }
     return output;