+++ injectable_generator/lib/lean_builder/lean_injectable_config_builder.dart
@@ -1,23 +1,18 @@
 import 'dart:convert';
-import 'dart:io';
 
 import 'package:code_builder/code_builder.dart';
 import 'package:collection/collection.dart';
 import 'package:dart_style/dart_style.dart';
-import 'package:glob/glob.dart';
-import 'package:glob/list_local_fs.dart';
 import 'package:injectable/injectable.dart';
 import 'package:injectable_generator/code_builder/allocator.dart';
 import 'package:injectable_generator/code_builder/library_builder.dart';
 import 'package:injectable_generator/lean_builder/resolvers/lean_importable_type_resolver.dart';
 import 'package:injectable_generator/models/dependency_config.dart';
 import 'package:injectable_generator/models/external_module_config.dart';
-import 'package:injectable_generator/models/importable_type.dart';
 import 'package:lean_builder/builder.dart';
 import 'package:lean_builder/element.dart';
 import 'package:lean_builder/type.dart';
 
-import 'package:injectable_generator/utils.dart';
 import 'package:recase/recase.dart';
 
 import '../resolver_utils.dart';
@@ -50,7 +45,6 @@ class InjectableConfigGenerator
         .literalValue
         .cast<String>();
 
-    final usesNullSafety = annotation.getBool('usesNullSafety')!.value;
     final isMicroPackage = annotation.getBool('_isMicroPackage')!.value;
     final usesConstructorCallback = annotation
         .getBool('usesConstructorCallback')!
@@ -70,16 +64,11 @@ class InjectableConfigGenerator
             .map((e) => e.getString('name')?.value) ??
         {};
 
-    final includeMicroPackages = annotation
-        .getBool("includeMicroPackages")!
-        .value;
-
     final generateAccessors = annotation.getBool("generateAccessors")!.value;
 
-    final allowMultipleRegistrations =
-        annotation.getBool('allowMultipleRegistrations')!.value;
-
-    final rootDir = annotation.getString('rootDir')?.value;
+    final allowMultipleRegistrations = annotation
+        .getBool('allowMultipleRegistrations')!
+        .value;
 
     final jsonData = <Map>[];
 
@@ -140,27 +129,6 @@ class InjectableConfigGenerator
     final microPackagesModules = microPackageModulesBefore.union(
       microPackageModulesAfter,
     );
-    if (!isMicroPackage && includeMicroPackages) {
-      final glob = Glob('**.module.dart', recursive: true);
-      final filesStream = glob.list(root: rootDir);
-
-      await for (final match in filesStream) {
-        final commentAnnotation = File(match.path).readAsLinesSync().first;
-        if (commentAnnotation.contains('@GeneratedMicroModule')) {
-          final segments = commentAnnotation.split(';');
-          if (segments.length == 3) {
-            final externalModule = ExternalModuleConfig(
-              ImportableType(name: segments[1], import: segments[2]),
-            );
-            if (!microPackagesModules.any(
-              (e) => externalModule.module == e.module,
-            )) {
-              microPackageModulesBefore.add(externalModule);
-            }
-          }
-        }
-      }
-    }
 
     final ignoreTypesInPackages =
         annotation
@@ -232,7 +200,6 @@ class InjectableConfigGenerator
     final emitter = DartEmitter(
       allocator: HashedAllocator(),
       orderDirectives: true,
-      useNullSafetySyntax: usesNullSafety,
     );
 
     final output = DartFormatter(
@@ -240,10 +207,8 @@ class InjectableConfigGenerator
     ).format(generatedLib.accept(emitter).toString());
 
     if (isMicroPackage) {
-      final outputUri = buildStep.asset.uriWithExtension('.module.dart');
       return buildStep.writeAsString(
         [
-          '//@GeneratedMicroModule;${capitalize(buildStep.asset.packageName!.pascalCase)}PackageModule;$outputUri',
           defaultFileHeader,
           output,
         ].join('\n'),